pyproject.toml
[tool.poetry]
name = "fastapi-hls"
version = "0.1.0"
description = ""
authors = ["Danny-Dasilva <yahchayildasilva@gmail.com>"]
license = "GPL3"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
click = "^8.1.3"
ffmpeg-python = "^0.2.0"
numpy = "^1.22.4"
uvicorn = "^0.17.6"
python-multipart = "^0.0.5"
aiofiles = "^0.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
